Refs #2992 Site Search KABOOM, Refs #49 Implementing Site Search tracking & reporting in Piwik core!
 * New Admin UI to customize, for each site, wheter site search is enabled. Also options to set default values to use.
 * New Reports: Searches, Searches with no result, Search categories, Top Pages Following a Search
 * to track "No result keyword" users will have to tag their site with a JS call, or add a new parameter to the search result page &search_count=X (X being zero for no result searches)
 * Reports works with Row evolution, PDF/HTML reports, Piwik Mobile
 * idaction_url is now NULLable because, Site Search records a page with idaction_name == Keyword, and idaction_url == NULL. This ensures that the Site Searches don't create "Page URL Not defined" records. 
 * updates to Tracker JS API, new function trackSiteSearch, also added in PHP tracker
 * New fields in log_visit to track searches
 * new segment, "searches" which can be used to select visitors who did a search ie. searches>0 or those who searched a lot, ie searches>10
 
 TODO: 
  * commmit integration test, TESTING, DOCS, FAQ, release, and a nice Prayer to the universe and the stars, hoping that I can code a major new feature without any bug...
  * It would be awesome to have compatiblity with Transitions so we can see, for a given site search, what are the starting pages and Destination pages
  
 Thank you for your patience Timo, 
 and thank in advance everyone for your help Testing this new feature!
 


git-svn-id: http://dev.piwik.org/svn/trunk@7190 59fd770c-687e-43c8-a1e3-f5a4ff64c105

+++ core/Piwik.php
@@ -2453,6 +2453,7 @@ class Piwik
 
 				return true;
 			} catch(Exception $e) {
+//				echo $sql . ' ---- ' .  $e->getMessage();
 				if(!Zend_Registry::get('db')->isErrNo($e, '1148'))
 				{
 					Piwik::log("LOAD DATA INFILE failed... Error was:" . $e->getMessage());
@@ -2498,16 +2499,23 @@ class Piwik
 
 				self::createCSVFile($filePath, $fileSpec, $values);
 
+				if(!file_exists($filePath))
+				{
+					throw new Exception("File $filePath could not be created.");
+				}
 				$rc = self::createTableFromCSVFile($tableName, $fields, $filePath, $fileSpec);
 				if($rc)
 				{
 					unlink($filePath);
 					return true;
 				}
-
-				throw new Exception('unknown cause');
+				else
+				{
+					throw new Exception('unknown cause');
+				}
 
 			} catch(Exception $e) {
+//				echo "LOAD DATA INFILE failed or not supported, falling back to normal INSERTs... Error was:" . $e->getMessage();
 				Piwik::log("LOAD DATA INFILE failed or not supported, falling back to normal INSERTs... Error was:" . $e->getMessage());
 			}
 		}
